requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML and NLP
transformers==4.36.0
torch==2.1.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
spacy==3.7.2

# PII Detection
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
nameparser==1.1.3
phonenumbers==8.13.25

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography==41.0.7

# Payment
stripe==7.8.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
sentry-sdk[fastapi]==1.38.0

# Utilities
httpx==0.25.2
pyyaml==6.0.1
python-dotenv==1.0.0
email-validator==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0